Fix inconsistent single and double quotes
It seems like single quotes and double quotes are being used
interchangably in the code. I could not find any underlying pattern
behind the use of single quotes or double quotes so I switched
everything over to double quotes. This provides the code base with a
higher degree of consistency.